LayoutTests/fast/wcss/wap-input-format-expected.txt

Assume the user input the following "AaBbCcDdEeFfGg123456!@#$%^,.;" to each of the input elements
{-wap-input-format:"*A";} only takes upper letters, symbols, punctuations
PASS domValueOf("star_upA") is "ABCDEFG!@#$%^,.;"
PASS visibleValueOf("star_upA") is "ABCDEFG!@#$%^,.;"
{-wap-input-format:"*a";} only takes lower letters, symbols, punctuations
PASS domValueOf("star_lowa") is "abcdefg!@#$%^,.;"
PASS visibleValueOf("star_lowa") is "abcdefg!@#$%^,.;"
{-wap-input-format:"*N";} only takes numbers 
PASS domValueOf("star_upN") is "123456"
PASS visibleValueOf("star_upN") is "123456"
{-wap-input-format:"*n";} only takes numbers , symbals and punctuations 
PASS domValueOf("star_lown") is "123456!@#$%^,.;"
PASS visibleValueOf("star_lown") is "123456!@#$%^,.;"
{-wap-input-format:"*X";} takes upper letters,numbers,symbals and punctuations 
PASS domValueOf("star_upX") is "ABCDEFG123456!@#$%^,.;"
PASS visibleValueOf("star_upX") is "ABCDEFG123456!@#$%^,.;"
{-wap-input-format:"*x";} takes lower letters,numbers,symbals and punctuations 
PASS domValueOf("star_lowx") is "abcdefg123456!@#$%^,.;"
PASS visibleValueOf("star_lowx") is "abcdefg123456!@#$%^,.;"
{-wap-input-format:"*M";} takes any character 
PASS domValueOf("star_upM") is "AaBbCcDdEeFfGg123456!@#$%^,.;"
PASS visibleValueOf("star_upM") is "AaBbCcDdEeFfGg123456!@#$%^,.;"
{-wap-input-format:"*m";} takes any character 
PASS domValueOf("star_lowm") is "AaBbCcDdEeFfGg123456!@#$%^,.;"
PASS visibleValueOf("star_lowm") is "AaBbCcDdEeFfGg123456!@#$%^,.;"
{-wap-input-format:"3N";} takes 3 numbers 
PASS domValueOf("three_N") is "123"
PASS visibleValueOf("three_N") is "123"
        
